pyproject.toml
[tool.poetry]
name = "python-tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "python_tracker"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.97.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
pyyaml = "^6.0"
sqlmodel = "^0.0.8"
psycopg2-binary = "^2.9.6"
alembic = "^1.11.1"
asyncpg = "^0.28.0"
python-multipart = "^0.0.6"
passlib = {extras = ["bccrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
